+++ src/Console/Command/ExtensionEnableCommand.php
@@ -19,15 +19,37 @@ use InvalidArgumentException;
 use LogicException;
 use ReflectionException;
 use Setup;
+use Symfony\Component\Console\Command\Command as SymfonyCommand;
+use Symfony\Component\Console\Input\InputArgument;
+use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 
-class ExtensionEnableCommand
+class ExtensionEnableCommand extends SymfonyCommand
 {
     public const DEFAULT_COMMAND = 'extension:enable';
     public const USAGE = self::DEFAULT_COMMAND . ' [filename] [classname]';
 
+    protected static $defaultName = self::DEFAULT_COMMAND;
+
     /** @var OutputInterface */
     private $output;
+
+    protected function configure(): void
+    {
+        $this
+            ->addArgument('filename', InputArgument::REQUIRED)
+            ->addArgument('classname', InputArgument::REQUIRED);
+    }
+
+    public function execute(InputInterface $input, OutputInterface $output): int
+    {
+        $filename = $input->getArgument('filename');
+        $classname = $input->getArgument('classname');
+
+        $this($output, $filename, $classname);
+
+        return 0;
+    }
 
     public function __invoke(OutputInterface $output, $filename, $classname): void
     {